Refactor ContestTable
