--- app/components/ContestButton/messages.js
@@ -1,21 +1,0 @@
-/*
- * ContestButton Messages
- *
- * This contains all the text for the ContestButton component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.ContestButton.header',
-    defaultMessage: 'This is the ContestButton component !',
-  },
-  enter: {
-    id: 'app.components.ContestButton.enter',
-    defaultMessage: 'ENTER',
-  },
-  join: {
-    id: 'app.components.ContestButton.join',
-    defaultMessage: 'JOIN',
-  },
-});
