--- app/components/ContestStatus/messages.js
@@ -1,25 +1,0 @@
-/*
- * ContestStatus Messages
- *
- * This contains all the text for the ContestStatus component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.ContestStatus.header',
-    defaultMessage: 'This is the ContestStatus component !',
-  },
-  enrolmentTo: {
-    id: 'app.components.ContestStatus.enrolmentTo',
-    defaultMessage: 'Enrolment to',
-  },
-  activeTo: {
-    id: 'app.components.ContestStatus.activeTo',
-    defaultMessage: 'Active to',
-  },
-  endedOn: {
-    id: 'app.components.ContestStatus.endedOn',
-    defaultMessage: 'Ended on',
-  },
-});
