+++ app/components/ContestsTable/Button/index.js
@@ -1,14 +1,7 @@
-/**
-*
-* ContestButton
-*
-*/
-
 import React from 'react';
 import moment from 'moment';
 import { Button } from 'material-ui/Button';
 import { Link } from 'react-router';
-// import styled from 'styled-components';
 
 import { FormattedMessage } from 'react-intl';
 import messages from './messages';
@@ -18,7 +11,7 @@ function ContestButton(props) {
   const enrolEndTime = moment(start).add(signupDuration, 'seconds');
   const buttonText = props.time.isAfter(enrolEndTime) ? <FormattedMessage {...messages.enter} />
     : <FormattedMessage {...messages.join} />;
-  const path = props.time.isAfter(enrolEndTime) ? `contests/${id}` : `contests/${id}/join`;
+  const path = props.time.isAfter(enrolEndTime) ? `/contests/${id}` : `/contests/${id}/join`;
   return (
     <Link to={path}>
       <Button raised primary>
