+++ app/components/ContestsTable/Button/messages.js
@@ -1,0 +1,12 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  enter: {
+    id: 'app.components.ContestTable.Button.enter',
+    defaultMessage: 'ENTER',
+  },
+  join: {
+    id: 'app.components.ContestTable.Button.join',
+    defaultMessage: 'JOIN',
+  },
+});
