+++ app/components/ContestsTable/Row/messages.js
@@ -1,0 +1,24 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  name: {
+    id: 'app.components.ContestsTable.Row.name',
+    defaultMessage: 'Name',
+  },
+  description: {
+    id: 'app.components.ContestsTable.Row.description',
+    defaultMessage: 'Description',
+  },
+  owners: {
+    id: 'app.components.ContestsTable.Row.owners',
+    defaultMessage: 'Owners',
+  },
+  startDate: {
+    id: 'app.components.ContestsTable.Row.startDate',
+    defaultMessage: 'Start date',
+  },
+  endDate: {
+    id: 'app.components.ContestsTable.Row.endDate',
+    defaultMessage: 'End date',
+  },
+});
