+++ app/components/ContestsTable/Status/messages.js
@@ -1,0 +1,16 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  enrolmentTo: {
+    id: 'app.components.ContestTable.Status.enrolmentTo',
+    defaultMessage: 'Enrolment to',
+  },
+  activeTo: {
+    id: 'app.components.ContestTable.Status.activeTo',
+    defaultMessage: 'Active to',
+  },
+  endedOn: {
+    id: 'app.components.ContestTable.Status.endedOn',
+    defaultMessage: 'Ended on',
+  },
+});
