--- app/components/ContestsTable/messages.js
@@ -1,33 +1,0 @@
-/*
- * ContestsTable Messages
- *
- * This contains all the text for the ContestsTable component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.ContestsTable.header',
-    defaultMessage: 'This is the ContestsTable component !',
-  },
-  name: {
-    id: 'app.components.ContestsTable.name',
-    defaultMessage: 'Name',
-  },
-  description: {
-    id: 'app.components.ContestsTable.description',
-    defaultMessage: 'Description',
-  },
-  owners: {
-    id: 'app.components.ContestsTable.owners',
-    defaultMessage: 'Owners',
-  },
-  startDate: {
-    id: 'app.components.ContestsTable.startDate',
-    defaultMessage: 'Start date',
-  },
-  endDate: {
-    id: 'app.components.ContestsTable.endDate',
-    defaultMessage: 'End date',
-  },
-});
